+++ R/withSpinner.R
@@ -28,7 +28,7 @@
 #' size of the image is used. Ignored if not using `image`.
 #' @param hide.ui By default, while an output is recalculating, the output UI is hidden and the spinner is visible instead.
 #' Setting `hide.ui = FALSE` will result in the spinner showing up on top of the previous output UI.
-#' @param caption Caption to display below the spinner or image. Ignored if `type` is set to 1 for css reasons.
+#' @param caption Caption to display below the spinner or image. Ignored if `type` is 1.
 #' @seealso [showSpinner()], [hideSpinner()], [pageSpinner()]
 #' @examples
 #' if (interactive()) {
@@ -62,7 +62,7 @@ withSpinner <- function(
   image.width = getOption("spinner.image.width"),
   image.height = getOption("spinner.image.height"),
   hide.ui = getOption("spinner.hide.ui", default = TRUE),
-  caption = NULL
+  caption = getOption("spinner.caption")
 ) {
 
   if (!inherits(ui_element, "shiny.tag") && !inherits(ui_element, "shiny.tag.list")) {