+++ NEWS.md
@@ -7,7 +7,7 @@
 - Fix bug where `shiny::tagList()` would throw an error (#73)
 - Add more informative error messages (#69)
 - Replace `addResourcePath()` with `htmlDependency()` which is more robust (#75)
-- Add `caption` argument to `withSpinner`
+- Add `caption` argument to allow showing a message under the spinner (#28)
 
 # shinycssloaders 1.0.0 (2020-07-28)